+++ substrate/frontier/src/interfaces.ts
@@ -1,19 +1,22 @@
-
-export interface ChainContext {
-    _chain: {
-        runtime: {
-            getEventTypeHash(name: string): string
-            getCallTypeHash(name: string): string
-        }
-    }
+interface Runtime {
+    getEventTypeHash(name: string): string
+    getCallTypeHash(name: string): string
 }
+
 
 export interface Event {
     name: string
     args: any
+    block: {
+        _runtime: Runtime
+    }
 }
+
 
 export interface Call {
     name: string
     args: any
+    block: {
+        _runtime: Runtime
+    }
 }