+++ substrate/substrate-data/src/interfaces/data-raw.ts
@@ -7,6 +7,6 @@ export interface RawBlock extends BlockData {
     validators?: AccountId[]
     session?: Bytes
     runtime?: Runtime
-    runtimeOfPreviousBlock?: Runtime
+    runtimeOfPrevBlock?: Runtime
     feeMultiplier?: Bytes
 }