+++ substrate/substrate-data/src/parser.ts
@@ -1,16 +1,16 @@
-import * as raw from '@subsquid/substrate-data-raw'
-import {HashAndHeight, Prev, runtimeVersionEquals} from '@subsquid/substrate-data-raw'
+import {HashAndHeight, Prev, Rpc} from '@subsquid/substrate-data-raw'
 import {OldSpecsBundle, OldTypesBundle} from '@subsquid/substrate-metadata'
-import {assertNotNull, groupBy, last} from '@subsquid/util-internal'
+import {annotateAsyncError, annotateSyncError, assertNotNull, groupBy, last} from '@subsquid/util-internal'
 import {RequestsTracker} from '@subsquid/util-internal-ingest-tools'
 import {RangeRequestList} from '@subsquid/util-internal-range'
 import assert from 'assert'
 import {Block, Bytes, DataRequest} from './interfaces/data'
 import {RawBlock} from './interfaces/data-raw'
 import {BlockParser} from './parsing/block'
+import {setEmittedContractAddress, setEthereumTransact, setEvmLog, setGearProgramId} from './parsing/extension'
 import {supportsFeeCalc} from './parsing/fee'
 import {AccountId} from './parsing/validator'
-import {Runtime} from './runtime'
+import {RuntimeTracker} from './runtime-tracker'
 
 
 const STORAGE = {
@@ -23,20 +23,31 @@ const STORAGE = {
 export class Parser {
     private requests: RequestsTracker<DataRequest>
     private prevValidators = new Prev<{session: Bytes, validators: AccountId[]}>()
-    private prevRuntime = new Prev<Runtime>()
+    private runtimeTracker: RuntimeTracker<RawBlock>
 
     constructor(
-        private rpc: raw.Rpc,
+        private rpc: Rpc,
         requests: RangeRequestList<DataRequest>,
-        private typesBundle?: OldTypesBundle | OldSpecsBundle
+        typesBundle?: OldTypesBundle | OldSpecsBundle
     ) {
         this.requests = new RequestsTracker(requests)
+
+        this.runtimeTracker = new RuntimeTracker<RawBlock>(
+            block => ({
+                height: block.height,
+                hash: block.hash,
+                parentHash: block.block.block.header.parentHash
+            }),
+            block => assertNotNull(block.runtimeVersion),
+            this.rpc,
+            typesBundle
+        )
     }
 
     async parse(blocks: RawBlock[]): Promise<Block[]> {
         if (blocks.length == 0) return []
 
-        await this.setRuntime(blocks)
+        await this.runtimeTracker.setRuntime(blocks)
 
         let result = []
 
@@ -54,7 +65,7 @@ export class Parser {
             }
 
             for (let block of batch.blocks) {
-                let parsed = parseRawBlock(block, batch.request)
+                let parsed = this.parseBlock(block, batch.request)
                 result.push(parsed)
             }
         }
@@ -62,62 +73,54 @@ export class Parser {
         return result
     }
 
-    private async setRuntime(blocks: RawBlock[]): Promise<void> {
-        if (blocks.length == 0) return
+    @annotateSyncError(getRefCtx)
+    private parseBlock(rawBlock: RawBlock, options?: DataRequest): Block {
+        let parser = new BlockParser(rawBlock, options)
 
-        let parentParentHeight = Math.max(0, blocks[0].height - 2)
-        let prev = this.prevRuntime.getItem(parentParentHeight)
-        if (prev == null) {
-            prev = await this.fetchRuntime(await this.getParent(getParent(blocks[0])))
+        let block = new Block(
+            assertNotNull(rawBlock.runtime),
+            assertNotNull(rawBlock.runtimeOfPrevBlock),
+            parser.header()
+        )
+
+        if (options?.blockTimestamp && parser.timestamp()) {
+            block.header.timestamp = parser.timestamp()
         }
 
-        if (runtimeVersionEquals(prev.value, assertNotNull(blocks[0].runtimeVersion)) || prev.height == parentParentHeight) {
-            blocks[0].runtimeOfPreviousBlock = prev.value
-        } else {
-            prev = await this.fetchRuntime(await this.getParent(getParent(blocks[0])))
-            blocks[0].runtimeOfPreviousBlock = prev.value
+        if (options?.blockValidator && parser.validator()) {
+            block.header.validator = parser.validator()
         }
 
-        for (let i = 0; i < blocks.length; i++) {
-            let block = blocks[i]
-            if (runtimeVersionEquals(prev.value, assertNotNull(block.runtimeVersion))) {
-                block.runtime = prev.value
-                prev = {
-                    height: block.height,
-                    value: prev.value
+        if (options?.calls) {
+            block.extrinsics = parser.extrinsics()?.map(item => item.extrinsic)
+            block.calls = parser.calls()
+            if (block.calls) {
+                for (let call of block.calls!) {
+                    setEthereumTransact(call)
                 }
-            } else if (prev.height == getParent(block).height) {
-                block.runtime = prev.value
+            }
+        }
+
+        if (options?.events) {
+            block.events = parser.events()
+            if (block.events) {
+                for (let event of block.events!) {
+                    setEvmLog(event)
+                    setEmittedContractAddress(event)
+                    setGearProgramId(event)
+                }
+            }
+        }
+
+        if (options?.extrinsicFee) {
+            if (parser.runtime.hasEvent('TransactionPayment.TransactionFeePaid')) {
+                parser.setExtrinsicFeesFromPaidEvent()
             } else {
-                prev = await this.fetchRuntime(getParent(block))
-                block.runtime = prev.value
-            }
-            if (i > 0) {
-                assert(blocks[i-1].height + 1 == block.height)
-                block.runtimeOfPreviousBlock = blocks[i-1].runtime
+                parser.calcExtrinsicFees()
             }
         }
-    }
 
-    private async getParent(ref: HashAndHeight): Promise<HashAndHeight> {
-        if (ref.height == 0) return ref
-        let header = await this.rpc.getBlockHeader(ref.hash)
-        return {
-            height: ref.height - 1,
-            hash: header.parentHash
-        }
-    }
-
-    private async fetchRuntime(
-        ref: HashAndHeight
-    ): Promise<{height: number, value: Runtime}> {
-        let [runtimeVersion, metadata] = await Promise.all([
-            this.rpc.getRuntimeVersion(ref.hash),
-            this.rpc.getMetadata(ref.hash)
-        ])
-        let runtime = new Runtime(runtimeVersion, metadata, this.typesBundle)
-        this.prevRuntime.set(ref.height, runtime)
-        return {height: ref.height, value: runtime}
+        return block
     }
 
     private async setValidators(blocks: RawBlock[]): Promise<void> {
@@ -147,6 +150,7 @@ export class Parser {
         }
     }
 
+    @annotateAsyncError(getRefCtx)
     private async fetchValidators(block: RawBlock): Promise<{session: Bytes, validators: AccountId[]}> {
         let [session, data] = await Promise.all([
             block.session ? Promise.resolve(block.session) : this.rpc.getStorage(block.hash, STORAGE.session),
@@ -178,46 +182,9 @@ export class Parser {
 }
 
 
-function getParent(block: RawBlock): HashAndHeight {
-    if (block.height == 0) return block
+function getRefCtx(ref: HashAndHeight) {
     return {
-        height: block.height - 1,
-        hash: block.block.block.header.parentHash
+        blockHeight: ref.height,
+        blockHash: ref.hash
     }
-}
-
-
-function parseRawBlock(rawBlock: RawBlock, options?: DataRequest): Block {
-    let parser = new BlockParser(rawBlock, options)
-
-    let block: Block = {
-        header: parser.header()
-    }
-
-    if (options?.blockTimestamp && parser.timestamp()) {
-        block.header.timestamp = parser.timestamp()
-    }
-
-    if (options?.blockValidator && parser.validator()) {
-        block.header.validator = parser.validator()
-    }
-
-    if (options?.calls) {
-        block.extrinsics = parser.extrinsics()?.map(item => item.extrinsic)
-        block.calls = parser.calls()
-    }
-
-    if (options?.events) {
-        block.events = parser.events()
-    }
-
-    if (options?.extrinsicFee) {
-        if (parser.runtime.hasEvent('TransactionPayment.TransactionFeePaid')) {
-            parser.setExtrinsicFeesFromPaidEvent()
-        } else {
-            parser.calcExtrinsicFees()
-        }
-    }
-
-    return block
 }