+++ substrate/frontier/CHANGELOG.md
@@ -1,6 +1,20 @@
 # Change Log - @subsquid/frontier
 
-This log was last generated on Wed, 19 Jul 2023 16:48:14 GMT and should not be manually modified.
+This log was last generated on Tue, 01 Aug 2023 16:40:12 GMT and should not be manually modified.
+
+## 3.0.0
+Tue, 01 Aug 2023 16:40:12 GMT
+
+### Breaking changes
+
+- adapt to per block runtime tracking in substrate processor
+
+## 2.0.1
+Thu, 27 Jul 2023 13:21:04 GMT
+
+### Patches
+
+- bump `ethers` peer to `^6.6.5`
 
 ## 2.0.0
 Wed, 19 Jul 2023 16:48:14 GMT