+++ substrate/ink-typegen/CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log - @subsquid/ink-typegen
 
-This log was last generated on Wed, 19 Jul 2023 16:48:14 GMT and should not be manually modified.
+This log was last generated on Tue, 01 Aug 2023 16:40:12 GMT and should not be manually modified.
+
+## 2.0.1
+Tue, 01 Aug 2023 16:40:12 GMT
+
+_Version update only_
 
 ## 2.0.0
 Wed, 19 Jul 2023 16:48:14 GMT