+++ substrate/substrate-data/CHANGELOG.md
@@ -1,6 +1,40 @@
 # Change Log - @subsquid/substrate-data
 
-This log was last generated on Wed, 19 Jul 2023 16:48:14 GMT and should not be manually modified.
+This log was last generated on Fri, 04 Aug 2023 07:25:15 GMT and should not be manually modified.
+
+## 1.2.0
+Fri, 04 Aug 2023 07:25:15 GMT
+
+### Minor changes
+
+- extract and expose `RuntimeTracker`
+- re-export `Rpc` from raw package
+
+## 1.1.0
+Wed, 02 Aug 2023 22:59:19 GMT
+
+### Minor changes
+
+- support indexing of special events and calls
+
+## 1.0.0
+Tue, 01 Aug 2023 16:40:12 GMT
+
+### Breaking changes
+
+- migrate to per block runtime tracking
+
+### Patches
+
+- annotate block parsing errors with `.blockHeight` and `.blockHash`
+- fix extrinsic fee calculation
+
+## 0.0.1
+Thu, 27 Jul 2023 13:21:04 GMT
+
+### Patches
+
+- fix block validator parsing
 
 ## 0.0.0
 Wed, 19 Jul 2023 16:48:14 GMT