+++ substrate/substrate-typegen/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @subsquid/substrate-typegen
 
-This log was last generated on Wed, 19 Jul 2023 16:48:14 GMT and should not be manually modified.
+This log was last generated on Tue, 01 Aug 2023 16:40:12 GMT and should not be manually modified.
+
+## 5.0.0
+Tue, 01 Aug 2023 16:40:12 GMT
+
+### Breaking changes
+
+- adapt to per block runtime tracking in substrate processor
 
 ## 4.0.0
 Wed, 19 Jul 2023 16:48:14 GMT